pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nrxn1-sv-analysis"
version = "0.1.0"
description = "NRXN1 Structural Variant Analysis Pipeline"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nagi Debbah", email = "your.email@example.com"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = ["bioinformatics", "genomics", "cnv", "structural-variants", "nrxn1", "neurexin"]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.4.0",
    "scipy>=1.8.0",
    "scikit-learn>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "plotly>=5.8.0",
    "biopython>=1.79",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.1.0",
    "pytest-cov>=3.0.0",
    "black>=22.3.0",
    "flake8>=4.0.0",
    "mypy>=0.960",
]
ml = [
    "xgboost>=1.6.0",
    "lightgbm>=3.3.0",
    "shap>=0.41.0",
]
viz = [
    "dash>=2.4.0",
    "streamlit>=1.8.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]
cloud = [
    "boto3>=1.24.0",
    "google-cloud-storage>=2.3.0",
]
all = [
    "nrxn1-sv-analysis[dev,ml,viz,cloud]",
]

[project.scripts]
nrxn1-analysis = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/nrxn1-sv-analysis"
Documentation = "https://github.com/yourusername/nrxn1-sv-analysis#readme"
Repository = "https://github.com/yourusername/nrxn1-sv-analysis"
Issues = "https://github.com/yourusername/nrxn1-sv-analysis/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
